requirements.txt
# Core NLP
emoji>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
scipy>=1.11.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.4.0
imbalanced-learn>=0.12.0

# Explainability
shap>=0.45.0

# data 
kagglehub

google-genai
python-dotenv
openai  